Fix stuck selection rects and small refactor
